+++ src/lib.rs
@@ -2,5 +2,6 @@
 
 pub mod action;
 pub mod combinator;
+pub mod instant;
 pub mod parser;
 pub mod range;